update: 'modified' component separators update in real-time
Before, dynamic components such as the 'Buffer Modified' built-in
component would not update its separators to 'disappear' when the file
included no modifications. This commit fixes this by converting the Lua
function call to determine if a separator should be included into
statusline syntax.

This was previously solved using autocmds in the 'attach_component'
function, however, this meant that a user had to specify a specific
'detach_on' condition for the specific component. This has now been
solved by having a conditional in the builtin components which returns
an empty string.