fix: trial `surround_whitespace()` created instead of double whitespace.
There's a mysterious 'bug' with Vim's statusline in which items
encapsulated inside `%{}` syntax require two spaces before any content
to produce a single space in the statusline. This commit introduces a
fix which uses the function `surround_whitespace()` to encapsulate
content using a Unicode space, which produces the expected result in
which a space is equal to strictly one space.

I'm not sure if this is the best way to go about creating built-in
functions in `builtins.lua` - it may be best to automatically run the
`surround_whitespace()` call elsewhere in nvim-stratus so that if users
create custom functions they do not have to worry about this edge-case.